# API.yaml
...
swagger: "2.0"
info:
  description: "Calculates minimum path between points depending on user generated\
    \ map input."
  version: "1.0.0"
  title: "Team 2 Pathfinding API"
host: "ec2-54-200-185-55.us-west-2.compute.amazonaws.com:8080"
basePath: "/v2"
schemes:
- "http"
paths:
...
  /id={problem_id}/Obstacles:
    get:
      tags:
      - "Obstacles"
      summary: "Obstacles"
      description: "Returns a list of all of the obstacles in the problem. This can\
        \ be an empty list.\n"
      operationId: "get_obstacles"
      produces:
      - "application/json"
      - "application/xml"
      parameters:
      - name: "problem_id"
        in: "path"
        description: "The id of the problem being manipulated"
        required: true
        type: "integer"
        format: "int32"
      responses:
        200:
          description: "An array of obstacles"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Obstacle"
        403:
          description: "ProblemID busy"
        default:
          description: "Unexpected error"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "swagger_server.controllers.obstacles_controller"
    post:
      tags:
      - "Obstacles"
      summary: "Add a new obstacle to the list"
      description: ""
      operationId: "add_obstacle"
      consumes:
      - "application/json"
      - "application/xml"
      produces:
      - "application/json"
      - "application/xml"
      parameters:
      - name: "problem_id"
        in: "path"
        description: "The id of the problem being manipulated"
        required: true
        type: "integer"
        format: "int32"
      - in: "body"
        name: "obstacle"
        description: "Obstacle object that needs to be added to the list."
        required: true
        schema:
          $ref: "#/definitions/Obstacle"
      responses:
        200:
          description: "The id of the new obstacle generated"
          schema:
            type: "integer"
            format: "int32"
        403:
          description: "ProblemID busy"
        405:
          description: "Invalid input"
      x-swagger-router-controller: "swagger_server.controllers.obstacles_controller"
  /id={problem_id}/Obstacles/obstacle_id={obstacle_id}:
    get:
      tags:
      - "Obstacles"
      summary: "Obstacles"
      description: "Returns an obstacle\n"
      operationId: "get_obstacle"
      produces:
      - "application/json"
      - "application/xml"
      parameters:
      - name: "problem_id"
        in: "path"
        description: "The id of the problem being manipulated"
        required: true
        type: "integer"
        format: "int32"
      - name: "obstacle_id"
        in: "path"
        description: "The id of the obstacle to be updated."
        required: true
        type: "integer"
        format: "int32"
      responses:
        200:
          description: "An obstacle"
          schema:
            $ref: "#/definitions/Obstacle"
        403:
          description: "ProblemID busy"
        default:
          description: "Unexpected error"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "swagger_server.controllers.obstacles_controller"
    put:
      tags:
      - "Obstacles"
      summary: "Update an existing obstacle"
      description: ""
      operationId: "update_obstacle"
      consumes:
      - "application/json"
      - "application/xml"
      produces:
      - "application/json"
      - "application/xml"
      parameters:
      - name: "problem_id"
        in: "path"
        description: "The id of the problem being manipulated"
        required: true
        type: "integer"
        format: "int32"
      - name: "obstacle_id"
        in: "path"
        description: "The id of the obstacle to be updated."
        required: true
        type: "integer"
        format: "int32"
      - in: "body"
        name: "updated_obstacle"
        description: "Obstacle object that needs to be added to the list."
        required: false
        schema:
          $ref: "#/definitions/Obstacle"
      responses:
        400:
          description: "Invalid ID supplied"
        403:
          description: "ProblemID busy"
        404:
          description: "Obstacle not found"
        405:
          description: "Validation exception"
      x-swagger-router-controller: "swagger_server.controllers.obstacles_controller"
    delete:
      tags:
      - "Obstacles"
      summary: "Delete Obstacle"
      description: "This removes the obstacle by the given ID"
      operationId: "delete_obstacle"
      produces:
      - "application/json"
      - "application/xml"
      parameters:
      - name: "problem_id"
        in: "path"
        description: "The id of the problem being manipulated"
        required: true
        type: "integer"
        format: "int32"
      - name: "obstacle_id"
        in: "path"
        description: "The ID of the Obstacle that needs to be deleted."
        required: true
        type: "integer"
        format: "int32"
      responses:
        400:
          description: "Invalid ID supplied"
        403:
          description: "ProblemID busy"
        404:
          description: "Obstacle not found"
      x-swagger-router-controller: "swagger_server.controllers.obstacles_controller"
...
  /id={problem_id}/Path:
    get:
      tags:
      - "Path"
      summary: "Path"
      description: "Returns a description of the path from the robot's current location\
        \ to the goal."
      operationId: "get_path"
      produces:
      - "application/json"
      - "application/xml"
      parameters:
      - name: "problem_id"
        in: "path"
        description: "The id of the problem being manipulated"
        required: true
        type: "integer"
        format: "int32"
      responses:
        200:
          description: "Array of Coordinates from start to finish."
          schema:
            $ref: "#/definitions/Path"
        403:
          description: "ProblemID busy"
        default:
          description: "Unexpected error"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "swagger_server.controllers.path_controller"
definitions:
  Coordinate:
    type: "object"
    properties:
      longitude:
        type: "number"
        format: "double"
        description: "the longitude of the object."
      latitude:
        type: "number"
        format: "double"
        description: "the latitude of the object."
  Rectangle:
    type: "object"
    properties:
      coordinates:
        $ref: "#/definitions/Coordinate"
      width:
        type: "number"
        format: "double"
        description: "The width of the rectangle"
      height:
        type: "number"
        format: "double"
        description: "The height of the rectangle"
  Obstacle:
    type: "object"
    properties:
      obstacle_id:
        type: "integer"
        format: "int32"
        description: "Unique identifier representing a specific obstacle in the list."
      obstacle_info:
        $ref: "#/definitions/Rectangle"
...
  Path:
    type: "object"
    properties:
      path_cost:
        type: "number"
        format: "double"
        description: "The total cost to travel the path"
      coordinates:
        type: "array"
        description: "The path travelled"
        items:
          $ref: "#/definitions/Coordinate"
...
  Error:
    type: "object"
    properties:
      code:
        type: "integer"
        format: "int32"
      message:
        type: "string"
      fields:
        type: "string"